Core: editor controller.

+++ pages/apps/generic/editor/editorController.ts
@@ -1,5 +1,7 @@
 import EventEmitter from "eventemitter3";
-import { GraphDB } from "~/server/api/graph/graphDB";
+import GraphAPI from "../../../../utils/GraphAPI";
+import type { IQwieryNode } from "~/utils";
+import { Utils } from "../../../../utils/utils/lib/utils";
 
 /*
  * MVC controller between the graph viewer and the backend (via GraphAPI).
@@ -7,7 +9,10 @@ import { GraphDB } from "~/server/api/graph/graphDB";
 export default class EditorController extends EventEmitter {
 	public commitChanges: boolean = true;
 
-	async deleteNode(node) {
+	async deleteNode(node: IQwieryNode) {
+		if (!node || Utils.isEmpty(node.id)) {
+			throw new Error("Nothing to delete.");
+		}
 		if (node && this.commitChanges) {
 			await GraphAPI.deleteNode(node.id);
 		}
@@ -16,9 +21,12 @@ export default class EditorController extends EventEmitter {
 
 	async createNode() {}
 
-	async updateNode(node) {
-		if (node && this.commitChanges) {
-			await GraphAPI.updateNode(node.data(), node.id(), node.data("labels"));
+	async updateNode(node: IQwieryNode) {
+		if (!node || Utils.isEmpty(node.id)) {
+			throw new Error("Nothing to update.");
+		}
+		if (this.commitChanges) {
+			await GraphAPI.updateNode(node);
 		}
 		this.emit("nodePropertiesUpdated", node);
 	}
